rename package data to model
