+++ src/main/java/au/com/mehdi/replacer/ui/InputForm.java
@@ -1,6 +1,6 @@
 package au.com.mehdi.replacer.ui;
 
-import au.com.mehdi.replacer.data.ReplaceData;
+import au.com.mehdi.replacer.model.ReplaceData;
 
 /**
  * Created by mehdi on 5/29/16.
